+++ ttnn/cpp/ttnn/graph/graph_query_op_constraints.hpp
@@ -29,17 +29,12 @@ namespace detail {
 inline Tensor extract_output_tensor(const Tensor& result) { return result; }
 
 // conv2d output
-template <typename... Args>
-Tensor extract_output_tensor(const std::tuple<Tensor, Args...>& result) {
-    return std::get<0>(result);
-}
-
-template <typename... Args1, typename... Args2, typename... Args3>
+template <typename... Args1, typename... Args2>
 Tensor extract_output_tensor(const std::variant<
                              ttnn::Tensor,
                              std::tuple<ttnn::Tensor, Args1...>,
                              std::tuple<ttnn::Tensor, Args2...>,
-                             std::tuple<ttnn::Tensor, Args3...>>& result) {
+                             std::tuple<ttnn::Tensor, Args1..., Args2...>>& result) {
     return std::visit<Tensor>(
         [](auto&& arg) {
             using T = std::decay_t<decltype(arg)>;
